add icons and css
