--- packages/theme-patternfly-org/pages/404.css
@@ -1,19 +1,0 @@
-#ws-404-search-wrapper {
-  position: relative;
-  max-width: 400px;
-  margin: 0 auto;
-}
-
-.ws-404-secondary-actions {
-  text-align: left;
-  flex-wrap: nowrap !important;
-}
-
-.ws-404-card-header {
-  padding: 1.5rem;
-}
-
-.ws-404-empty-state-card {
-  flex: 1;
-  margin: var(--pf-global--spacer--md);
-}
