+++ packages/theme-patternfly-org/scripts/webpack/webpack.client.config.js
@@ -87,7 +87,7 @@ const clientConfig = async (env, argv) => {
             {
               loader: MiniCssExtractPlugin.loader,
               options: {
-                hmr: !isProd
+                hmr: false
               }
             },
             {
